# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: ERP Barrio API
  version: 0.3.8
  description: Contrato inicial del API del MVP ERP.
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    BranchCreateRequest:
      type: object
      required: [code, name, address]
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        is_active:
          type: boolean
          default: true
    BranchUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        is_active:
          type: boolean
    BranchResponse:
      type: object
      required: [id, code, name, address, is_active]
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        address:
          type: string
        is_active:
          type: boolean
    BranchListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BranchResponse'
...
    DocumentTypeCreateRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        requires_expiry:
          type: boolean
          default: true
        is_active:
          type: boolean
          default: true
    DocumentTypeUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        requires_expiry:
          type: boolean
        is_active:
          type: boolean
    DocumentTypeResponse:
      type: object
      required: [id, code, name, requires_expiry, is_active]
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        requires_expiry:
          type: boolean
        is_active:
          type: boolean
    DocumentTypeListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentTypeResponse'
...
paths:
...
  /branches:
    get:
      tags: [branches]
      summary: Lista sucursales de la organización
      operationId: getBranches
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Listado de sucursales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchListResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [branches]
      summary: Crea una sucursal del ERP
      operationId: postBranches
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchCreateRequest'
      responses:
        '201':
          description: Sucursal creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Código de sucursal duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /branches/{branch_id}:
    parameters:
      - in: path
        name: branch_id
        required: true
        schema:
          type: string
    get:
      tags: [branches]
      summary: Obtiene una sucursal por identificador
      operationId: getBranchById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sucursal encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sucursal no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [branches]
      summary: Actualiza campos de una sucursal
      operationId: patchBranchById
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchUpdateRequest'
      responses:
        '200':
          description: Sucursal actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sucursal no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [branches]
      summary: Elimina una sucursal del ERP
      operationId: deleteBranchById
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Sucursal eliminada
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sucursal no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /document-types:
    get:
      tags: [document-types]
      summary: Lista tipos documentales del ERP
      operationId: getDocumentTypes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Listado de tipos documentales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeListResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [document-types]
      summary: Crea un tipo documental del ERP
      operationId: postDocumentTypes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentTypeCreateRequest'
      responses:
        '201':
          description: Tipo documental creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Código de tipo documental duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /document-types/{document_type_id}:
    parameters:
      - in: path
        name: document_type_id
        required: true
        schema:
          type: string
    get:
      tags: [document-types]
      summary: Obtiene un tipo documental por identificador
      operationId: getDocumentTypeById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tipo documental encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tipo documental no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [document-types]
      summary: Actualiza campos de un tipo documental
      operationId: patchDocumentTypeById
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentTypeUpdateRequest'
      responses:
        '200':
          description: Tipo documental actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tipo documental no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [document-types]
      summary: Elimina un tipo documental del ERP
      operationId: deleteDocumentTypeById
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Tipo documental eliminado
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tipo documental no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...